Correção de problemas na listagem das tasks com filtro

+++ app/src/main/java/br/edu/ifsp/dmo1/gerenciadortarefasmvvm/ui/listener/TaskClickListener.kt
@@ -1,5 +1,5 @@
 package br.edu.ifsp.dmo1.gerenciadortarefasmvvm.ui.listener
 
 interface TaskClickListener {
-    fun clickDone(position: Int)
+    fun clickDone(position: Long)
 }
